\name{coso_sta_LLZ}
\alias{coso_sta_LLZ}
\title{Coso Station File}
\description{Coso Station Location file, 1989-1999
}
\usage{data(coso_sta_LLZ)
}
\format{Name, Lat, Lon, Z
}
\source{
 Personal Files
}
\references{
  Wu, H. and J. M. Lees (1996). Attenuation Structure of Coso Geothermal
  Area, California, from P Wave Pulse Widths, Bull. Seismol. Soc. Am.,
  86, 1574-1590.

Lees, J. M. (1998), Multiplet analysis at Coso Geothermal,Bull. Seismol. Soc. Am. 88(5) 1127-1143.
  
}
\keyword{datasets}
