\name{gMAP}
\alias{gMAP}
\title{Generic Map Button
}
\description{Generic Map Button
}
\usage{
gMAP(nh, g, idev = 3)
}
\arguments{
  \item{nh}{RSEIS structure 
}
  \item{g}{parameters used in swig
}
  \item{idev}{device for plotting (not used)
}
}
\details{This is a button used internally in swig
}
\value{Graphical Side Effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{swig
}
\examples{
\dontrun{

buts = "gMAP"
swig(GH, PADDLAB = buts )


}


}
\keyword{misc}
