\name{imageINFLUENCE}
\alias{imageINFLUENCE}
\title{Image Influence of stations
}
\description{Plot contours/image of Influence scores.
}
\usage{
imageINFLUENCE(B, sta, proj)
}
\arguments{
  \item{B}{Pseudovalue list
}
  \item{sta}{station location list
}
  \item{proj}{projection list
}
}
\details{Following jackknife - plot results.
  this function is called by
  plotJACKLLZ.
}
\value{side effects
}
\references{

   Iversen, E. S., and J. M. Lees (1996), A statistical technique for validating velocity models, Bull. Seismol. Soc. Am. 86(6), 1853-1862.

}
\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{plotJACKLLZ
}

\keyword{misc}
