\name{lastPIX}
\alias{lastPIX}
\alias{editPIX}

\title{Last Pix
}
\description{
  RSEIS Button: Restore Last WPX file from
  memory. Function is used internally in swig.
}
\usage{
lastPIX(nh, g)
editPIX(nh, g)

}
\arguments{
  \item{nh}{GH list from RSEIS
}
  \item{g}{parameters from swig
}
}

\value{
  New WPX list attached to g
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\keyword{misc}
