\name{PostREQquake}
\alias{PostREQquake}
\title{Post Processing on EQrquake
}
\description{Post Processing on EQrquake
}
\usage{
PostREQquake(XQ, proj)
}
\arguments{
  \item{XQ}{List of Earthquakes
}
  \item{proj}{projection list
}
}
\details{Following event locations, plot.
}
\value{graphical side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\keyword{misc}
